parse release from fetch function
